chore: refactor repo discovery paths
